requirements.txt
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.127.0
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
langchain>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.1
colorama>=0.4.6
rich>=13.7.1
